requirements.txt
Flask
tensorflow-cpu
keras
numpy
gspread
google-auth
h5py
gunicorn
pillow
gdown